+++ Lib/fontbakery/specifications/adobe_fonts.py
@@ -50,7 +50,7 @@ expected_check_ids = [
     'com.google.fonts/check/whitespace_ink',
     'com.google.fonts/check/whitespace_widths',
     'com.google.fonts/check/required_tables',
-    'com.google.fonts/check/053',  # Are there unwanted tables?
+    'com.google.fonts/check/unwanted_tables',
     'com.google.fonts/check/057',  # Name table entries should not contain line-breaks.
     'com.google.fonts/check/058',  # Glyph names are all valid?
     'com.google.fonts/check/059',  # Font contains unique glyph names?